+++ spinta/dimensions/prefix/helpers.py
@@ -1,9 +1,10 @@
 from typing import Any
 from typing import Dict
+from typing import Union
 
 from spinta import commands
-from spinta.components import Component
 from spinta.components import Context
+from spinta.datasets.components import Dataset
 from spinta.dimensions.prefix.components import UriPrefix
 from spinta.manifests.components import Manifest
 from spinta.nodes import get_node
@@ -12,7 +13,7 @@ from spinta.nodes import get_node
 def load_prefixes(
     context: Context,
     manifest: Manifest,
-    parent: Component,
+    node: Union[Manifest, Dataset],
     prefixes: Dict[str, Dict[str, Any]],
 ) -> Dict[str, UriPrefix]:
     config = context.get('config')
@@ -24,13 +25,12 @@ def load_prefixes(
             data['eid'],
             data,
             group='dimensions',
-            parent=parent,
+            parent=node,
         )
         prefix.eid = data['eid']
         prefix.type = data['type']
-        prefix.parent = parent
-        prefix.parent = manifest
-        commands.load(context, prefix, data, parent=parent)
+        prefix.parent = node
+        commands.load(context, prefix, data, parent=node)
         loaded[name] = prefix
     return loaded